Refactor language tag lookup to reduce code duplication
